chore: move validate_bearer_token to metrics file
